Consolidated the tests in one runner
