+++ tests/beta.c
@@ -1,8 +1,4 @@
-#include <check.h>
-
-#include "../ge.h"
-
-#define NOP2_OPCODE 0x07
+#include "tests.h"
 
 START_TEST(execute_nop)
 {
@@ -38,7 +34,7 @@ START_TEST(execute_nop)
     ck_assert_uint_eq(g.rPO, 0x0002);
 }
 
-Suite * init_suite(void)
+Suite *init_beta_suite(void)
 {
     Suite *s;
     TCase *tc_core;
@@ -54,19 +50,3 @@ Suite * init_suite(void)
 
     return s;
 }
-
-int main(void)
-{
-    int number_failed;
-    Suite *s;
-    SRunner *sr;
-
-    s = init_suite();
-    sr = srunner_create(s);
-
-    srunner_run_all(sr, CK_NORMAL);
-    number_failed = srunner_ntests_failed(sr);
-    srunner_free(sr);
-    return (number_failed == 0) ? 0 : -1;
-}
-
